+++ src/pages/DetailPage/DetailPage.tsx
@@ -17,20 +17,13 @@ import {
   DetailInfo2,
 } from './styles';
 
-import {
-  doc,
-  setDoc,
-  getDoc,
-  updateDoc,
-  increment,
-  DocumentData,
-} from 'firebase/firestore';
+import { doc, setDoc, getDoc, updateDoc, increment } from 'firebase/firestore';
 import { useQuery } from 'react-query';
-import { useParams, useNavigate } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { fetchSpotDetailData } from '../../apis/publicAPI';
 import { FetchedStayDataType } from '../../types/apiDataTypes';
 import Loader from '../../components/Loader/Loader';
-import React, { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { db } from '../../apis/firebase';
 import RestaurantInfo from '../../components/Recommendation/Info/RestaurantInfo';
 import StayInfo from '../../components/Recommendation/Info/StayInfo';
@@ -41,14 +34,12 @@ import noimg from '../../assets/noimg.avif';
 
 import MobileCommunication from '../../components/Selection/mobile/MobileCommunication';
 
-import SpotLiked from '../../components/Liked/SpotLiked';
+import PlaceLiked from '../../components/Liked/PlaceLiked';
 import SideInfoMap from '../../components/Map/SideInfoMap';
 import DetailFooter from '../../components/Footer/DetailFooter';
 import { useMediaQuery } from 'react-responsive';
 const DetailPage = () => {
   const param = useParams();
-  const navigate = useNavigate();
-  // const [likeData, setLikeData] = useState<DocumentData | undefined>();
 
   const { data: spotDetailData, isLoading: isLoadingSpot } = useQuery(
     ['spot_detail', param],
@@ -121,7 +112,11 @@ const DetailPage = () => {
                     {spotDetailData.addr1.split(' ', 2)}
                   </DetailTextArr>
                   <DetailImojiBox>
-                    <SpotLiked spotDetailData={spotDetailData} />
+                    <PlaceLiked
+                      spotDetailData={spotDetailData}
+                      stayDetailData={undefined}
+                      restaurantDetailData={undefined}
+                    />
                     {/* <p>{likeData !== undefined ? likeData.likeCnt : 0}</p> */}
                   </DetailImojiBox>
                 </DetailTextBox>
